py/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.0
beautifulsoup4>=4.12.2
oauthlib>=3.2.2
google-auth>=2.23.3
python-dotenv>=1.0.0
email-validator>=2.1.0
yandex-oauth==1.1.2
mysql-connector-python>=8.0.33
cryptography==41.0.3
PyJWT==2.8.0
pyyaml==6.0.1
PySocks==1.7.1
python-dateutil
sqlalchemy>=2.0.0
itsdangerous>=2.1.0
python-multipart>=0.0.6
crawl4ai
openai
anthropic
